refactorizando el codigo de los post
