+++ src/lib/posts.js
@@ -1,0 +1,5 @@
+import firebase from './secret.js';
+
+const db = firebase.firestore();
+
+export const getAllPost = () => db.collection('post');
